Define trivial  generators for requests and response
